+++ packages/codepen/index.js
@@ -1,4 +1,3 @@
-import "/src/style.css";
-import "@lil-el/ui/css";
+import "./src/assets/index.css";
 
 export * from "/src/components/code/index.js";