+++ packages/codepen/tailwind.config.js
@@ -1,0 +1,7 @@
+export default {
+  content: ["./node_modules/@lil-el/**/*.{js}"],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+};