+++ packages/markdown/index.js
@@ -1,4 +1,4 @@
-import "./src/style.css";
+import "./src/assets/index.css";
 
 import { defineAsyncComponent } from "vue";
 export const markdown = defineAsyncComponent(() => import("./src/components/markdown/index.vue"));