+++ packages/markdown/src/components/index.js
@@ -1,5 +1,4 @@
 import { defineAsyncComponent } from "vue";
-import "@lil-el/codepen/css";
 
 export const helloworld = defineAsyncComponent(() => import("./helloworld/index.vue"))