+++ packages/markdown/src/main.js
@@ -1,6 +1,6 @@
 import { createApp } from "vue";
+
 import "./style.css";
-import "@lil-el/codepen/css";
 import App from "./App.vue";
 
 createApp(App).mount("#app");